djangoapp/requirements/base.txt
pytz==2018.5  # https://github.com/stub42/pytz
python-slugify==1.2.6  # https://github.com/un33k/python-slugify
Pillow==5.3.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
redis==3.2.1  # https://github.com/antirez/redis
importlib-metadata==4.13.0
celery==4.4.2  # https://github.com/celery/celery
lxml==4.2.5  # https://github.com/lxml/lxml
tornado==5.1.1  # https://github.com/tornadoweb/tornado
kombu>=4.6.8<4.7  # https://github.com/celery/kombu
psycopg2-binary==2.7.5  # https://github.com/psycopg/psycopg2
vine==1.3.0

# Django
# -----------------------------------------------------------------------------
django==2.0.9  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.37.1  # https://github.com/pennersr/django-allauth
django-rest-auth==0.9.3  # https://github.com/Tivix/django-rest-auth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.10.0  # https://github.com/niwinz/django-redis
python-postmark==0.5.0  # https://github.com/themartorana/python-postmark
django-dotenv==1.4.2  # https://github.com/jpadilla/django-dotenv
django-celery-tracker==1.3.0  # https://github.com/chris-allen/django-celery-tracker
django-filter==2.0.0  # https://github.com/carltongibson/django-filter
django-reversion==3.0.7
django-summernote==0.8.20.0 # https://github.com/summernote/django-summernote

# Django REST Framework
djangorestframework==3.8.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
drf-extensions==0.4.0  # https://github.com/chibisov/drf-extensions
djangorestframework-jwt==1.11.0
cryptography==2.6.1
PyJWT==1.7.1
django-cors-headers==2.5.2
djangorestframework-xml==1.4.0

# Pusher
pusher==2.1.3
pusher_push_notifications==1.1.0

# Zip Code Lookup
uszipcode==1.0.1

# Celery Beat Status
django-celerybeat-status==0.0.10

# DRF Swagger Generator
drf-yasg==1.17.1

# Make requests to 3rd party api
requests==2.23.0

# Soft Delete Functionality
soft-delete==0.2.2


# Custom
python-Levenshtein
